requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
Werkzeug==3.0.1
marshmallow==3.20.2
email-validator==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
cryptography==41.0.7
